backend/requirements.txt
fastapi
uvicorn
SQLAlchemy[asyncio]
asyncmy
passlib[bcrypt]
python-dotenv
cryptography>=3.4.7
PyJWT
